Restructure the code with modularity
