+++ index.css
@@ -16,26 +16,23 @@
 }
 
 .nav-items > ul > li {
-  padding: 20px;
+  padding: 10px;
 
 } 
+
+
 .card{
   width: 200px;
   border: 1px solid black;
-  padding: 5px;
-  margin: auto;
-  margin-bottom: 10px;
-  display: flex;
-  flex-wrap: wrap;
-  flex-grow: inherit;
+  padding: 10px;
+  margin: 10px;
 }
 
-.card > img{
+.card > img {
   width: 100%;
 }
 
-.restaurant-list{
+.restaurant-list {
   display: flex;
   flex-wrap: wrap;
 }
-
