+++ index.html
@@ -14,6 +14,6 @@
     
 
     <!-- Injected external js file -->
-    <script type="module" src="App.js"></script>
+    <script type="module" src="/src/App.js"></script>
   </body>
 </html>
