+++ src/Components/Footer.jsx
@@ -1,0 +1,5 @@
+const FooterComponent = () =>{
+    return <h3>Footer Components</h3>
+  }
+
+  export default FooterComponent;
